requirements.txt

openai>=1.0.0
anthropic>=0.5.0
pyyaml>=6.0
pandas>=1.3.0
numpy>=1.20.0
tqdm>=4.62.0
python-dotenv>=0.19.0
rich>=10.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
ray>=2.0.0